--- data-logger/python/pose_utils/__init__.py
@@ -1,1 +1,0 @@
-from .pose_utils import getAllMotionPackets
