--- data-logger/python/pose_utils/pose_utils.py
@@ -1,48 +1,0 @@
-import TimestampedPacketMotionData_pb2
-import PoseSequenceLabel_pb2
-import TimestampedImage_pb2
-import Vector3dStamped_pb2
-def getAllMotionPackets(motion_data_folder: str, use_json: bool):
-   motion_packets = []
-   if use_json:
-      filepaths = [os.path.join(motion_data_folder, f) for f in os.listdir(motion_data_folder) if os.path.isfile(os.path.join(motion_data_folder, f)) and str.lower(os.path.splitext(f)[1])==".json"]
-      jsonstrings = [(open(path, 'r')).read() for path in filepaths]
-      for jsonstring in jsonstrings:
-         data = TimestampedPacketMotionData_pb2.TimestampedPacketMotionData()
-         google.protobuf.json_format.Parse(jsonstring, data)
-         motion_packets.append(data)
-   else:
-      filepaths = [os.path.join(motion_data_folder, f) for f in os.listdir(motion_data_folder) if os.path.isfile(os.path.join(motion_data_folder, f)) and str.lower(os.path.splitext(f)[1])==".pb"]
-      for filepath in filepaths:
-         try:
-            data = TimestampedPacketMotionData_pb2.TimestampedPacketMotionData()
-            f = open(filepath,'rb')
-            data.ParseFromString(f.read())
-            f.close()
-            motion_packets.append(data)
-         except:
-            print("Could not read udp file %s." %(filepath))
-            continue
-   return motion_packets
-def getAllImageFilePackets(image_data_folder: str, use_json: bool):
-   image_packets = []
-   if use_json:
-      filepaths = [os.path.join(image_data_folder, f) for f in os.listdir(image_data_folder) if os.path.isfile(os.path.join(image_data_folder, f)) and str.lower(os.path.splitext(f)[1])==".json"]
-      jsonstrings = [(open(path, 'r')).read() for path in filepaths]
-      for jsonstring in jsonstrings:
-         data = TimestampedImage_pb2.TimestampedImage()
-         google.protobuf.json_format.Parse(jsonstring, data)
-         image_packets.append(data)
-   else:
-      filepaths = [os.path.join(image_data_folder, f) for f in os.listdir(image_data_folder) if os.path.isfile(os.path.join(image_data_folder, f)) and str.lower(os.path.splitext(f)[1])==".pb"]
-      for filepath in filepaths:
-         try:
-            data = TimestampedImage_pb2.TimestampedImage()
-            f = open(filepath,'rb')
-            data.ParseFromString(f.read())
-            f.close()
-            image_packets.append(data)
-         except:
-            print("Could not read image data file %s." %(filepath))
-            continue
-   return image_packets
